model-Concept-AB/requirements.txt
numpy>=1.20
matplotlib>=3.3
pyyaml>=5.4
scipy>=1.6
jupyter
